Ensures Pods get restarted when config changes.
fixes #2

This patch includes two related changes:

1. The Pod has two new annotations that contain a checksum of each ConfigMap.
That way when the contents of a ConfigMap change, that will result in a change
to the Pod template, which will in turn cause the Deployment to replace any
current Pods with a new rollout.

2. The creation of kube resource objects is refactored to happen on a single
object. At the beginning of reconciliation, all resources get created once
up-front and are stored in that object for use throughout the reconciliation
workflow. This is especially helpful since resource creation involves rendering
at least one template, which can return an error; handling that case once
up-front makes it possible to then easily reference resources and their
attributes from other code. Specific to this change, it became necessary to
reference the ConfigMaps from multiple places. Rather than re-render templates
and re-handle error scenarios, it made sense to create the ConfigMaps once and
reference them from a source of truth.