Locate an external CA using the image.config.openshift.io API
OpenShift uses the image.config.openshift.io API to store external
registry information so it can be used in the cluster.  This PR will
look at the field 'additionalTrustedCA' on the image.config.openshift.io
singleton to get the ConfigMap that hold the external registry's cert.
The operator will make a copy of this configmap in the local namespace
and mount it into the Cincinnati deployment.  Changes in the configmap
will results in Cincinnati being restarted, so certs this also supports
cert rotation.