Simplify the Cincinnati Operator API
Remove the CertConfigMapKey field and instead look for a hardcoded key,
'cincinnati-registry', in the openshift-config ConfigMap.  Having a
hardcoded value won't affect usability and it allows us to safely locate
the CA Cert.

The patch is a good segue into the future where we can potentially bundle
all the certs found in the openshift-config ConfigMap and mount them into
Cincinnati.