pkg/apis/cincinnati/v1beta1: Explicitly require registry and repository
Users must point the update service at a repository so the graph
builder can scrape for primary release metadata.

Also require replicas.  It's possible that we could make this optional
going forward, with the operator supplying a reasonable default.  But
there's no operator-side code for defaults today.

Also require graphDataImage.  It's possible that we could make this
optional going forward, with the operator disabling the graph-data
plugins in the policy engine when it is not set.  But there's no
operator-side code for that today, and
docs/graph-data-init-container.md documents how to build a graph-data
image pretty easily today.

Also require the spec, but leave status optional.  This allows users
to upload a manifest, while leaving the status empty for the
controller to fill in later.

The new metadata docstring is based on wording from [1].

I've dropped some 'omitempty' from struct properties, because Go's
encoding/json, etc., do not support omitempty for structs [2].  And
there's no need to set omitempty for required properties anyway.

I've pivoted some existing '// +optional' marker to
'+kubebuilder:validation:Optional', to make the consumer more obvious
(operator-sdk is in charge of invoking the kubebuilder logic).  Docs
for the kubebuilder markers are in [4].

Generated changes with:

  $ operator-sdk generate crds

using:

  $ operator-sdk version
  operator-sdk version: "v0.16.0", commit: "55f1446c5f472e7d8e308dcdf36d0d7fc44fc4fd", go version: "go1.13.8 linux/amd64"

[1]: https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.18/#pod-v1-core
[2]: https://github.com/golang/go/issues/11939
[3]: https://book.kubebuilder.io/reference/markers/crd-validation.html