pkg/apis/cincinnati/v1beta1: Drop redundant doc.go
We've had a package docstring in both doc.go and register.go since
v1alpha1 in 52a976598c (first code commit, 2020-04-03, #1).  But
there's no reason to have a doc.go declaration when register.go
already defines the same material.

Also add a trailing period to the docstring sentence and capitalize
the Cincinnati proper noun.  And remove the boilerplate comment,
because this is a hand-maintained file, even if changes will be rare.