docs: Set namespace for Cincinnati manifests
Or in the 'oc -n ... create ...' call that pushes them into the
cluster.  And also use the NAMESPACE variable when looking up the
policy engine's route.

We could add an environment variable for the Cincinnati manifest's
name, but I think 'example-name', which I'm pivoting to, is clear
enough in pattern-matching that we don't need a formal environment
variable.