build: Simplify Dockerfile and remove 'command'
The UBI image we're installing into should have HOME and all that
already set up for us, so no need to muck about with all that.  Also
drop the shell entrypoint wrapper and just call our binary directly.
Also drop the 'command' entry from the Deployment, because that's
already the entrypoint, so we don't need to specify it explicitly.