api/v1: Collapse 'registry' and 'repository' into 'releases'
"Registry" and "repository" are both generic names for portions of
container image pullspecs.  But graphDataImage is another pullspec in
the UpdateService spec.  Make the semantics more obvious, and remove
the artifical spliting, by collapsing the two properties into a single
'releases', named after the meaning of the value instead of the syntax
of the value.

I made the Go changes, and then grepped around for old-style
references to update.  Then I regenerated the CRD with:

  $ controller-gen crd:trivialVersions=true rbac:roleName=updateservice-operator webhook paths="./..." output:crd:artifacts:config=config/crd/bases

See the previous commits, especially 3990db3008 (controllers:
Consolidate kubebuilder declarations, 2020-11-12, #83), for more on
that controller-gen usage.