*: Use Git description for version information
We've been tracking a nominal operator version since 52a976598c (first
code commit, 2020-04-03), but this is one more thing we'd need to
remember to bump between releases, and it's hard to know what version
number we'd want to use before we are building a release.  Drop the
hard-coded strings from the operator to give us one less moving piece.
Also drop the SDK version, because the SDK is used to generate some
scaffolding, but it doesn't have much impact on the resulting operator
binary (no vendored libraries, etc.).

Because it's helpful for debugging when folks hand us operator logs,
I've shifted things around a bit to include the Git commit hash or
other description in the early logging.

The OLM bundles will remain versioned, but bundle version can be a
decision made at bundle-build time, pulling in whichever operator
image they want to pull in.