Operator bundle: changes to support operator bundling
Modifies Makefile and creates the config/manifests artifacts which
support bundling and is generated by the new recipe 'gen-for-bundle'.
The `scorecard` directory has been added back. It was originally
removed since we are not using operator-sdk Scorecard but it is
required to be presnet by `operator-sdk generate bundle`.

This PR assumes the following:

- Development will run the `gen-for-bundle` recipe and check the
results into the `cincinnati-operator` git repo.
- The individual running the official build will run the `bundle`
recipe in a temporary location and copy the resulting artifacts to
the proper location required by the official build tools.

The additional recipes added to the Makefile are described in detail
below.

'manifests` generates manifests, e.g. CRD, RBAC etc., and is run by
development when required.

'gen-for-bundle` generates kustomize bases and a kustomization.yaml
used to generate the operator bundle. This was originally part of the
`bundle` recipe as generated by the operator-sdk scaffolding. I have
broken it out under the assumption that we wll run this recipe and
check-in the resulting config/manifests directory and contents to the
git repo where they would be used to do an official build of the
operator bundle.

'bundle' generates the bundle and metadata as contained in a 'bundle'
directory and 'bundle.Dockerfile'.

'bundle-build' builds the bundle image locally and is intended for
use by developers.