document using shorter name for cincinnati cr
In case the user creates the Cincinnati CR with a longer name, there
might be a case where the route's hostname exceeds 63 chars. This will
result in the route not getting created, inturn crashing the operator.

Document using a shorter name, less than 64 characters for Cincinnati CR
to prevent the issue.