Add additionalPrinterColumns to the UpdateService crd
Add additionalPrinterColumns to the UpdateService custom resource
definition file. Previously running the "oc get -o wide ..." command
only displayed the name and the age of the specific running
UpdateService resources.

Display the policyEngineURI, releases, graphDataImage, and
the ReconcileCompleted condition status to provide more information
to a user about the UpdateService resource. Set the priority of these
additionalPrinterColumns to 1 to list them only when the "-o wide" flag
is provided.