add graph data image annotation to osus operand
we're tuning up a graph-data pod that refreshes after every
5 min and fetched the latest graph-data image digest which
we use to keep a track if the image tag was updated.

Logic:
1. we reconcile every 5 minutes
2. We tune up a graph-data pod which runs for 5 min
3. after the graph-data pod is in succeeded condition, the
reconcile loop will delete it to create a new one. This is
where we get the latest digest for the image.
4. we add this image as an annotation to the deployment.
So, everytime the annotation/image changes, it'll tune up
new operand pods.

We'll not be creating the graph-data pod if the image is
referenced by digest instead of tag.
We're creating just one graph-data pod for OSUS.