controllers: Create a NetworkPolicy for each UpdateService
Adds a basic NetworkPolicy that restricts the network communication
for the Pods delivering an UpdateService (there are no NetworkPolicies
yet for the operator itself, that will require further OLM work).  For
each UpdateService, the operator creates a NetworkPolicy that restrict
networking for all Pods related to that UpdateService.  The
NetworkPolicies are created before the Deployment, but because there
is no NetworkPolicy status to tell us whether the policy has been
picked up and applied, it may be a race for whether or not early Pods
are covered.  When an UpdateService is removed, the NetworkPolicy is
removed, along with all the other resources, via ownerReferences and
Kubernetes garbage-collection [1].

The rules are:

* Ingress from the Router, following the suggestion in [2].  Those
  docs currently contain a somewhat concerning caveat:

    (1) policy-group.network.openshift.io/ingress:"" label supports OVN-Kubernetes.

  I dunno what that means for folks who aren't running OVN.  Maybe we
  no longer support any networking that's not OVN, and we don't
  support folks installing custom networking?  Or maybe the docs are
  stale, and there isn't an OVN restriction.
* Egress to all targets and ports, to support graph-builder scraping
  registries.  Even if the scraping is routed via proxies.  In
  follow-up work, we can try to tighten this down.
* Egress to DNS.  This is redundant at the moment, because of the
  blanket egress allowance, but I'm coding it as an explicit entry so
  we can leave it unchanged as we tighten registry egress in the
  future.  Not clear to me how reliable "DNS will always come on 5353
  via pods in the openshift-dns namespace with a
  dns.operator.openshift.io/daemonset-dns=default label" will be in
  different environments, but that's what the NetworkPolicy folks are
  currently recommending in internal documents.

I'm not currently opening metrics to scraping, because we don't ship
any ServiceMonitors or PrometheusRule alerting or other metrics
consumers for the UpdateService Deployments.  I'm also not removing
the Service entries listing the metrics ports though, so not a
terribly consistent position.  There will likely be later clean up
work around metrics, although I'm not sure if it will tilt towards
making them useful or removing them.

The read access for all namespaces avoids [3]:

  2025-10-27T18:46:43.167059221Z W1027 18:46:43.167017       1 reflector.go:547] sigs.k8s.io/controller-runtime/pkg/cache/internal/informers.go:106: failed to list *v1.NetworkPolicy: networkpolicies.networking.k8s.io is forbidden: User "system:serviceaccount:openshift-update-service:updateservice-operator" cannot list resource "networkpolicies" in API group "networking.k8s.io" in the namespace "openshift-config"

where either the SDK we're using cannot distinguish between "some of
these resources we need to watch openshift-config for, but some just
need to be in the openshift-update-service namespace", or we're just
holding the SDK wrong.  I'm piling on to the existing pattern here,
and haven't dug in deeply to form new opinions on whether it's a good
pattern.

[1]: https://kubernetes.io/docs/concepts/architecture/garbage-collection/
[2]: https://docs.redhat.com/en/documentation/openshift_container_platform/4.20/html/network_security/network-policy#nw-networkpolicy-allow-from-router_about-network-policy
[3]: https://gcsweb-ci.apps.ci.l2s4.p1.openshiftapps.com/gcs/test-platform-results/pr-logs/pull/openshift_cincinnati-operator/243/pull-ci-openshift-cincinnati-operator-master-operator-e2e-new-ocp-published-graph-data/1982875203662778368/artifacts/operator-e2e-new-ocp-published-graph-data/e2e-test/artifacts/inspect/namespaces/openshift-update-service/pods/updateservice-operator-78c486b76-8jlpp/updateservice-operator/updateservice-operator/logs/current.log